pyproject.toml
[tool.poetry]
name = "lessons"
version = "0.1.0"
description = ""
authors = ["SvAS <sv4s.night@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
requests = "^2.32.3"
pytest = "^8.2.2"
pytest-cov = "^5.0.0"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.0"
black = "^24.4.2"
mypy = "^1.10.1"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
